Join Previous: Ignore commented lines
I don't want it to affect commented lines, because that's not any use for timing. I don't want to remove the comments though.
Now, if the previous line is commented, it goes to the line before it, and changes that instead.
If that line's commented, it goes to the line before that, and so on.

I tried to do it by just modifying subs[j] directly, but strangely, it didn't work.
So, I create the `previous` variable from subs[j], modify it, and put it right back into subs[j].
Obviously not ideal, but i tried the good thing and it didnt work so what am i supposed to do.

This commit also bumps the version number.