KaraTemplater: Add `nozerolen` modifier
Prevents execution of components for zero-duration syllables.
Name is open to discussion.

Potentially, this could be extended to lines, but I would propose
a second modifier for this, as there might be cases where one would
want to filter out zero-duration events on a `template char`
without disabling it for zero-length syls (e.g. punctuation)